webapp/requirements.txt
flask
python-dotenv
pymongo
flask-socketio
flask-login==0.6.2; python_version >= '3.7'
werkzeug==2.3.4; python_version >= '3.7'
pytest==7.4.0; python_version >= '3.7'
pytest-cov==3.0.0; python_version >= '3.7'
pymongo==4.3.3; python_version >= '3.7'